+++ CHANGELOG.md
@@ -16,8 +16,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - **CUMULUS-1853**
   - Added `@cumulus/integration-tests/collections.createCollection()`
   - Added `@cumulus/integration-tests/executions.findExecutionArn()`
-  - Added `@cumulus/integration-tests/executions.getCompletedExecution()`
-  - Added `@cumulus/integration-tests/executions.getFailedExecution()`
+  - Added `@cumulus/integration-tests/executions.getExecutionWithStatus()`
   - Added `@cumulus/integration-tests/granules.getCompletedGranule()`
   - Added `@cumulus/integration-tests/providers.createProvider()`
   - Added `@cumulus/integration-tests/rules.createOneTimeRule()`